config API: use error_errno() instead of error(), if we have errno
Improve the errors emitted e.g. when we run out of space on a
partition, and try to write a configuration file there. Before this
change we'd emit:

    $ git config --add foo.bar baz
    error: failed to write new configuration file /run/user/1001/linux/.git/config.lock

Now we'll emit a meaningful strerror():

    $ git config --add foo.bar baz
    error: failed to write new configuration file /run/user/1001/linux/.git/config.lock: No space left on device

This improves error handling code added in 480c9e521bd (short i/o: fix
config updates to use write_in_full, 2007-01-08). Let's make sure to
save away "errno" right after a syscall failure (or rather, the call
to wrapper.c, which is close enough).

In the preceding commits we refactored some of the code in a way that
introduced no functional changes, to implement this we'll now always
call a write_lock_in_full() wrapper for write_in_full() (whether
that's directly or indirectly).

I initially tried a "int *failure_errno" pattern, but it's much easier
and cleaner to simply have the error_errno() message emitted at the
point of failure instead.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>