config.c: use common templates for all error_errno() invocations
Rather than giving translators several messages that are similar to
translate, let's have one message for e.g. failing to call chmod(),
fstat() etc. on that lock file.

Let's also make it clear and consistent when we're talking about the
config file and when we're talking about the "new configuration file",
i.e. the lock file.

Finally, when we say "could not write config file %s" we should
mention both the lock file and the configuration file. If
commit_lock_file() fails we don't know which (or both) of the files
was involved in a failure, so let's mention both paths. Usually we'll
be failing on a rename() in that case.

A test relied on one of the the exact strings we're replacing, let's
update it, and include the ": " at the end to indicate that we're
emitting the strerror(). Let's also change "test_i18ngrep" to a plain
"grep" while we're at it.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>