tests: run the tests with GIT_NO_TEMPLATE_DIR=true (cleanup)
Clean up after the preceding commit where we started running the test
suite with GIT_NO_TEMPLATE_DIR=true.

Because of that preceding change we'll no longer need to pass
"--template=" in those cases were we wanted tests not to rely on our
templates, as it's now the default. We won't need the setting of
"TEST_CREATE_REPO_NO_TEMPLATE=1" before sourcing test-lib.sh either,
as that's now also the default.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>