run-command.c: name anonymous "children" struct & use variables
Name the "struct ... *children" introduced in
c553c72eed6 (run-command: add an asynchronous parallel child
processor, 2015-12-15) and use the new "struct parallel_process_child"
to avoid repeated mentions to "pp->children[i]", instead assign such
references once to a "ppc" variable, and use it to reference the child
we're currently processing.

This makes the control flow clearer, as it's now more obvious that the
"i" variable isn't being incremented after we find the child, and cuts
down on the verbosity of the long lines in the pre-image. That will
help to make subsequent commits smaller.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>