test-lib: add a "test_expect_todo", soft-deprecate "test_expect_failure"
The "test_expect_failure" function is almost never doing what we'd
like it to do. Introduce a "test_expect_todo" function intended to
replace it.

The problems with "test_expect_failure" are:

 * When it's used in combination with "test_must_fail" and
   "test_might_fail". It will hide segfaults or abort() failures, such
   as failures due to LSAN.

   This is because we do do the right thing in those helpers, but they
   themselves are expected to be used within "test_expect_success" and
   return 1. The "test_expect_failure" can't differentiate this from a
   "real" failure.

   To be fair we could change "test_{must,might}_fail" to
   appropriately carry forward the status code to work around this
   specific issue, but doing so would be a large semantic change in
   the current test suite.

 * It doesn't test what *does* happen, but just that it fails
   in *some* way. The problem is that today's failure might not be the
   same as tomorrow's.

   This isn't strictly speaking a problem in the sense that it *is*
   useful for cases where we know that something fails, but *how* it
   fails is unpredictable.

   The problem for our tests is that this almost never (or never)
   happens. We know we have a segfault, or that a "test_cmp" will fail
   or whatever.

For cases where we know *how* we failed we should just use the
semantics of "test_expect_failure", but the problem with that is that
we lose the self-documentation aspect of "test_expect_failure", both
in marking a test as "acts this way, but that's bad" to human readers,
and for machines in the case of the TAP output.

Let's introduce a "test_expect_todo" which addresses these
issues. Ideally we'd just replace "test_expect_failure" with it, but
doing so would require sweeping changes across the test suite. It:

 * Is like "test_expect_failure", except that "success" is emitted as
   a "not ok ... # TODO", just like "test_expect_failure".

 * Is unlike "test_expect_failure" in that anything except the test
   passing as "ok" (which can be e.g. an ending of
   test_{must,might}_fail) will be classified as a failure.

Let's convert a few "test_expect_failure" uses to the new
"test_expect_todo". E.g. the one in "t7815-grep-binary.sh" would
succeed with "test_expect_failure" if we added a "&& false" to it, but
"test_expect_todo" will correctly note that the failure is not the
particular one we expected (which is now marked with a
"test_must_fail").

For previous discussion on this topic see
https://lore.kernel.org/git/87tuhmk19c.fsf@evledraar.gmail.com/

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>