test-lib-functions: add a "test_todo", for use with "test_expect_todo"
Introduce a new "test_todo" function for use with the new
"test_expect_todo" function, and convert a test added in
aecad374ae7 (reflog-walk: don't segfault on non-commit sha1's in the
reflog, 2016-01-05) to use it.

A good thing about "test_expect_failure" is that it allowed for
writing the sort of tests added in aecad374ae7, where we declare what
we'd like to eventually get, but just mark that we're still failing to
get that.

For cases like the one-line "t7815-grep-binary.sh" test adjusted in
the preceding commit to use "test_expect_todo" it's obvious what we'd
eventually like to do, to remove the "test_must_fail" and convert
"test_expect_todo" to "test_expect_success".

But in the cases being adjusted here it would be a regression to
simply remove the "test_line_count", as the "expect" part of it is
what we'd eventually like to have pass.

By introducing a "test_todo" we can have our cake and eat it too. We
can declare both what we'll get now, as well as what we'd like to get
once the relevant feature is implemented (or a bug is fixed).

The "test_todo" function will always run the check for what we
eventually want, and BUG() out if it succeeds. We'll also BUG() out if
neither succeed.

This way we can ensure that unlike "test_expect_failure" blocks the
assertions aren't dead or broken code, as is the case for some in the
test suite. I.e. some are hopelessly broken and could never pass even
if a bug was fixed or a given feature was implemented, but we haven't
noticed because we've been treating all failures in
"test_expect_failure" equally.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>