test-lib-functions: add a "todo_test_cmp", for use with "test_expect_todo"
Introduce a new "todo_test_cmp" for use with the new
"test_expect_todo" function, and convert most of the tests added in
45bde58ef8f (grep: demonstrate bug with textconv attributes and
submodules, 2021-09-29) to use it.

This function is a trivial wrapper around doing the same with the
underlying "test_todo" function. Instead of the more verbose:

    test_todo test_cmp --want want --expect expect -- actual

We can now do:

    todo_test_cmp want expect actual

For cases like the one-line "t7815-grep-binary.sh" test adjusted in a
preceding commit to use "test_expect_todo" it's obvious what we'd
eventually like to do, to remove the "test_must_fail" and convert
"test_expect_todo" to "test_expect_success".

But in the cases being adjusted here it would be a regression to
simply remove the "test_cmp", as the "expect" part of it is what we'd
eventually like to have pass.

Hence the introduction of "todo_test_cmp": It's like "test_cmp", but
takes a mandatory additional first "want" argument, which documents
what we'd eventually want. The "expect" then becomes what we currently
get.

Since it uses "test_todo", this "test_cmp_todo" function will BUG()
out if "want" and "expect" are the same, and likewise if the "want" is
equivalent to "actual".

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>