tests: convert some "test_expect_failure" to "test_expect_todo"
Convert tests added in [1] and [2] to use "test_expect_todo" to more
specifically mark the failure condition, rather then the blank check
for failure given by "test_expect_failure".

For [1] there isn't an easy way to fit the "echo world" and "test_cmp"
into a "test_todo", and in any case we might not have that output once
we fix the bug noted in [1].

For [2] we'd need add "test_with_config" (which invokes both "git" and
"test-tool" to the list in "test_must_fail_acceptable", but adding a
file-specific function to "test-lib-functions.sh" would be a bit odd.

Let's instead expand the restrictive list added in
6a67c759489 (test-lib-functions: restrict test_must_fail usage,
2020-07-07). As shown in the commits that preceded it the list came
about because we wanted to exclude the likes of "cvs" and "p4" from
"test_might_fail".

It's a fair bet that if we use it with a test_* function that that
function is being appropriately used with it (i.e. invokes one of or
own programs). It's possible that without an exhaustive list we'll get
it wrong, but I also don't think it's worth it to maintain that
exhaustive list. Let's just allow all test_* names instead.

1. 90a6464b4ad (rerere: make sure it works even in a workdir attached
   to a young repository, 2011-03-10
2. 751d3b9415f (t1309: document cases where we would want early config
   not to die(), 2017-03-13)

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>