test-lib: add a BUG_ARITY, use it in test-lib-functions.sh
Add a BUG_ARITY for the common case of wanting to test the arity of
functions in test-lib-functions.sh.

When developing the subsequent patches I ran into some cases where I
got the arity wrong, and just getting an error like "1 param" wasn't
very useful. Let's always include the function name, and what arity we
expected in addition to what we got.

Let's also test that we emit the expected messages when we git it
wrong by adding the missing tests for this feature to t0000-basic.sh.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>