*.c: "return 0" in stub functions to appease SunCC
The SunCC compiler will complain about functions using BUG() or die()
that are non-void not having return statements.

I.e. it's not that it doesn't understand NORETURN, but that the check
for a function returning according to its signature has higher
priority than NORETURN analysis.

This fixes the following warnings:

    "t/helper/test-ref-store.c", line 260: warning: Function has no return statement : cmd_reflog_expire
    "t/helper/test-trace2.c", line 207: warning: Function has no return statement : ut_007bug

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>